Add tests, add Equation.Common module for tests
